Slepians.jl

This repo includes codes to compute discrete prolate spheroidal sequences, generalized prolate spheroidal sequences (unequal sampling), and solves the special case of d-dimensional concentration in the Cartesian plane where the function is restricted to the disc in the spectral domain.

These docs are under construction, but I will include the docstrings for the main functions here.

Index of functions

Slepians.dpss_tapers — Function

dpss_tapers(n,w,k,tap_or_egval)

Simply compute discrete prolate spheroidal sequence tapers, eigenvalues

...

Arguments

n::Int64: Length of the taper

nw::Float64: Time-bandwidth product

k::Int64: Number of tapers

tap_or_egval::Symbol = :tap: Either :tap, :egval, or :both

...

...

Outputs

vv::Vector{Float64}: The matrix of eigenvalues, if taporegval is set to :tap

dpss_eigval: Struct conaining the dpss tapers

...